+++ packages/titanium-sw-notifier/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.0.14 (2020-07-20)
+
+**Note:** Version bump only for package @leavittsoftware/titanium-sw-notifier
+
+
+
+
+
 ## [2.0.13](https://github.com/LeavittSoftware/titanium-elements/compare/@leavittsoftware/titanium-sw-notifier@2.0.12...@leavittsoftware/titanium-sw-notifier@2.0.13) (2020-07-06)
 
 **Note:** Version bump only for package @leavittsoftware/titanium-sw-notifier